refactor userController
