+++ config/default.json
@@ -1,4 +1,4 @@
 {
   "mongoURI": "mongodb+srv://mariawright:ijmagnm@news-explorer.qwfup.mongodb.net/myFirstDatabase?retryWrites=true&w=majority",
-  "jwtSecret": "theverybiggestsecret"
+  "JWT_SECRET": "theverybiggestsecret"
 }
