+++ controllers/userController.js
@@ -1,18 +1,16 @@
-const express = require('express');
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const config = require('config');
 
-const router = express.Router();
 const { check, validationResult } = require('express-validator');
-const auth = require('../middleware/auth');
 
+// const auth = require('../middleware/auth');
 const User = require('../models/user');
 
-// @route     GET /auth
+// @route     GET /users/me
 // @desc      Get logged in user (email, name)
 // @access    Private
-router.get('/', auth, async (req, res) => {
+const getUser = async (req, res) => {
   try {
     // req.user from auth middleware
     const user = await User.findById(req.user.id).select('-password');
@@ -21,16 +19,64 @@ router.get('/', auth, async (req, res) => {
     console.error(err.message);
     res.status(500).send('Server Error');
   }
-});
+};
 
-// @route     POST /auth
+// @route     POST /signup
+// @desc      Register a user
+// @access    Public
+const registerUser = async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    // change error
+    return res.status(400).json({ errors: errors.array() });
+  }
+
+  const { name, email, password } = req.body;
+
+  try {
+    let user = await User.findOne({ email });
+
+    if (user) {
+      // change error
+      return res.status(400).json({ msg: 'User already exists' });
+    }
+
+    user = new User({
+      name,
+      email,
+      password,
+    });
+
+    const salt = await bcrypt.genSaltSync(10);
+
+    user.password = await bcrypt.hash(password, salt);
+
+    await user.save();
+
+    const payload = {
+      user: {
+        id: user.id,
+      },
+    };
+
+    jwt.sign(payload, config.get('JWT_SECRET'), {
+      expiresIn: '7d',
+      // change error
+    }, (err, token) => {
+      if (err) throw (err);
+      res.json({ token });
+    });
+  } catch (err) {
+    // change error
+    console.error(err.message);
+    res.status(500).send('Server error');
+  }
+};
+
+// @route     POST /signin
 // @desc      Auth user & get token
 // @access    Public
-router.post('/', [
-  check('email', 'Please include a valid email').isEmail(),
-  check('password', 'Password is required').exists(),
-],
-async (req, res) => {
+const loginUser = async (req, res) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     // change error
@@ -68,6 +114,10 @@ async (req, res) => {
     console.error(err.message);
     res.status(500).send('Server Error');
   }
-});
+};
 
-module.exports = router;
+module.exports = {
+  getUser,
+  registerUser,
+  loginUser,
+};
