+++ middleware/auth.js
@@ -12,7 +12,7 @@ module.exports = function (req, res, next) {
 
   try {
     // Verify token and pull out payload
-    const decoded = jwt.verify(token, config.get('jwtSecret'));
+    const decoded = jwt.verify(token, config.get('JWT_SECRET'));
     // Gives access to token inside the route
     req.user = decoded.user;
     next();
